Refactor math methods to MathUtilities
